:truck: revert dir rename
